+++ src/N98/Magento/Application/ConfigurationLoader.php
@@ -2,13 +2,13 @@
 
 namespace N98\Magento\Application;
 
+use N98\Util\ArrayFunctions;
 use N98\Util\BinaryString;
 use N98\Util\OperatingSystem;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 use Symfony\Component\Yaml\Yaml;
-use N98\Util\ArrayFunctions;
 
 /**
  * Config consists of several parts which are merged.
@@ -94,7 +94,7 @@ class ConfigurationLoader
      * @param bool $isPharMode
      * @param OutputInterface $output
      */
-    public function __construct($config, $isPharMode, OutputInterface $output)
+    public function __construct(array $config, $isPharMode, OutputInterface $output)
     {
         $this->_initialConfig = $config;
         $this->_isPharMode = $isPharMode;
@@ -119,7 +119,7 @@ class ConfigurationLoader
 
     /**
      * @param string $magentoRootFolder
-     * @param bool   $loadExternalConfig
+     * @param bool $loadExternalConfig
      * @param string $magerunStopFileFolder
      */
     public function loadStageTwo($magentoRootFolder, $loadExternalConfig = true, $magerunStopFileFolder = '')
@@ -154,14 +154,14 @@ class ConfigurationLoader
      *
      * @return array
      */
-    protected function loadDistConfig($initConfig)
+    protected function loadDistConfig(array $initConfig)
     {
         if ($this->_distConfig == null) {
-            $this->_distConfig = Yaml::parse(__DIR__ . '/../../../../config.yaml');
+            $distConfigFilePath = __DIR__ . '/../../../../config.yaml';
+            $this->_distConfig = ConfigFile::createFromFile($distConfigFilePath)->toArray();
         }
-        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-            $this->_output->writeln('<debug>Load dist config</debug>');
-        }
+        $this->logDebug('Load dist config');
+
         $config = ArrayFunctions::mergeArrays($this->_distConfig, $initConfig);
 
         return $config;
@@ -174,7 +174,7 @@ class ConfigurationLoader
      *
      * @return array
      */
-    public function loadSystemConfig($config)
+    public function loadSystemConfig(array $config)
     {
         if ($this->_systemConfig == null) {
             if (OperatingSystem::isWindows()) {
@@ -184,11 +184,7 @@ class ConfigurationLoader
             }
 
             if ($systemWideConfigFile && file_exists($systemWideConfigFile)) {
-                if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-                    $this->_output->writeln(
-                        '<debug>Load system config <comment>' . $systemWideConfigFile . '</comment></debug>'
-                    );
-                }
+                $this->logDebug('Load system config <comment>' . $systemWideConfigFile . '</comment>');
                 $this->_systemConfig = Yaml::parse($systemWideConfigFile);
             } else {
                 $this->_systemConfig = array();
@@ -202,23 +198,25 @@ class ConfigurationLoader
     /**
      * Load config from all installed bundles
      *
-     * @param array  $config
+     * @param array $config
      * @param string $magentoRootFolder
      *
      * @return array
      */
-    public function loadPluginConfig($config, $magentoRootFolder)
+    public function loadPluginConfig(array $config, $magentoRootFolder)
     {
         if ($this->_pluginConfig == null) {
             $this->_pluginConfig = array();
             $moduleBaseFolders = array();
+            $customFilename = $this->_customConfigFilename;
+            $customName = pathinfo($customFilename, PATHINFO_FILENAME);
             if (OperatingSystem::isWindows()) {
-                $config['plugin']['folders'][] = getenv('WINDIR') . '/n98-magerun2/modules';
-                $config['plugin']['folders'][] = OperatingSystem::getHomeDir() . '/n98-magerun2/modules';
+                $config['plugin']['folders'][] = getenv('WINDIR') . '/' . $customName . '/modules';
+                $config['plugin']['folders'][] = OperatingSystem::getHomeDir() . '/' . $customName . '/modules';
             } else {
-                $config['plugin']['folders'][] = OperatingSystem::getHomeDir() . '/.n98-magerun2/modules';
+                $config['plugin']['folders'][] = OperatingSystem::getHomeDir() . '/.' . $customName . '/modules';
             }
-            $config['plugin']['folders'][] = $magentoRootFolder . '/lib/n98-magerun2/modules';
+            $config['plugin']['folders'][] = $magentoRootFolder . '/lib/' . $customName . '/modules';
             foreach ($config['plugin']['folders'] as $folder) {
                 if (is_dir($folder)) {
                     $moduleBaseFolders[] = $folder;
@@ -236,10 +234,11 @@ class ConfigurationLoader
                         ->depth(2)
                         ->followLinks()
                         ->ignoreUnreadableDirs(true)
-                        ->name($this->_customConfigFilename)
+                        ->name($customFilename)
                         ->in($this->getVendorDir());
 
-                    foreach ($finder as $file) { /* @var $file SplFileInfo */
+                    foreach ($finder as $file) {
+                        /* @var $file SplFileInfo */
                         $this->registerPluginConfigFile($magentoRootFolder, $file);
                     }
                 }
@@ -253,10 +252,11 @@ class ConfigurationLoader
                     ->depth(1)
                     ->followLinks()
                     ->ignoreUnreadableDirs(true)
-                    ->name($this->_customConfigFilename)
+                    ->name($customFilename)
                     ->in($moduleBaseFolders);
 
-                foreach ($finder as $file) { /* @var $file SplFileInfo */
+                foreach ($finder as $file) {
+                    /* @var $file SplFileInfo */
                     $this->registerPluginConfigFile($magentoRootFolder, $file);
                 }
             }
@@ -268,8 +268,8 @@ class ConfigurationLoader
     }
 
     /**
-     * @param string      $rawConfig
-     * @param string      $magentoRootFolder
+     * @param string $rawConfig
+     * @param string $magentoRootFolder
      * @param SplFileInfo $file [optional]
      *
      * @return string
@@ -278,37 +278,27 @@ class ConfigurationLoader
     {
         $replace = array(
             '%module%' => $file ? $file->getPath() : '',
-            '%root%'   => $magentoRootFolder,
+            '%root%' => $magentoRootFolder,
         );
 
         return str_replace(array_keys($replace), $replace, $rawConfig);
     }
 
-
     /**
      * Check if there is a user config file. ~/.n98-magerun.yaml
      *
-     * @param array  $config
+     * @param array $config
      * @param string $magentoRootFolder [optional]
      *
      * @return array
      */
-    public function loadUserConfig($config, $magentoRootFolder = null)
+    public function loadUserConfig(array $config, $magentoRootFolder = null)
     {
-        if ($this->_userConfig == null) {
+        if (null === $this->_userConfig) {
             $this->_userConfig = array();
-            $homeDirectory =  OperatingSystem::getHomeDir();
-            if (OperatingSystem::isWindows()) {
-                $personalConfigFile = $homeDirectory . DIRECTORY_SEPARATOR . $this->_customConfigFilename;
-            } else {
-                $personalConfigFile = $homeDirectory . DIRECTORY_SEPARATOR . '.' . $this->_customConfigFilename;
-            }
-
-            if ($homeDirectory && file_exists($personalConfigFile)) {
-                $userConfig = $this->applyVariables(\file_get_contents($personalConfigFile), $magentoRootFolder, null);
-                $this->_userConfig = Yaml::parse($userConfig);
-
-                return $config;
+            $locator = new ConfigLocator($this->_customConfigFilename, $magentoRootFolder);
+            if ($userConfigFile = $locator->getUserConfigFile()) {
+                $this->_userConfig = $userConfigFile->toArray();
             }
         }
 
@@ -326,37 +316,31 @@ class ConfigurationLoader
      *
      * @return array
      */
-    public function loadProjectConfig($magentoRootFolder, $magerunStopFileFolder, $config)
+    public function loadProjectConfig($magentoRootFolder, $magerunStopFileFolder, array $config)
     {
-        if ($this->_projectConfig == null) {
-            $this->_projectConfig = array();
-
-            $projectConfigFile = $magentoRootFolder . DIRECTORY_SEPARATOR . 'app/etc/' . $this->_customConfigFilename;
-            if ($projectConfigFile && file_exists($projectConfigFile)) {
-                $projectConfig = $this->applyVariables(
-                    \file_get_contents($projectConfigFile), $magentoRootFolder, null
-                );
-                $this->_projectConfig = Yaml::parse($projectConfig);
-            }
-
-            $stopFileConfigFile = $magerunStopFileFolder . DIRECTORY_SEPARATOR . '.' . $this->_customConfigFilename;
-            if (!empty($magerunStopFileFolder) && file_exists($stopFileConfigFile)) {
-                $projectConfig = $this->applyVariables(
-                    \file_get_contents($stopFileConfigFile), $magentoRootFolder, null
-                );
-                $this->_projectConfig = ArrayFunctions::mergeArrays($this->_projectConfig, Yaml::parse($projectConfig));
-            }
-
-            $config = ArrayFunctions::mergeArrays($config, $this->_projectConfig);
+        if (null !== $this->_projectConfig) {
+            return ArrayFunctions::mergeArrays($config, $this->_projectConfig);
         }
 
-        return $config;
+        $this->_projectConfig = array();
+
+        $locator = new ConfigLocator($this->_customConfigFilename, $magentoRootFolder);
+
+        if ($projectConfigFile = $locator->getProjectConfigFile()) {
+            $this->_projectConfig = $projectConfigFile->toArray();
+        }
+
+        if ($stopFileConfigFile = $locator->getStopFileConfigFile($magerunStopFileFolder)) {
+            $this->_projectConfig = $stopFileConfigFile->mergeArray($this->_projectConfig);
+        }
+
+        return ArrayFunctions::mergeArrays($config, $this->_projectConfig);
     }
 
     /**
      * Loads a plugin config file and merges it to plugin config
      *
-     * @param string       $magentoRootFolder
+     * @param string $magentoRootFolder
      * @param SplFileInfo $file
      */
     protected function registerPluginConfigFile($magentoRootFolder, $file)
@@ -371,19 +355,15 @@ class ConfigurationLoader
             }
 
             if ($path === false) {
-                $this->_output->writeln(sprintf("<error>Plugin config file broken link '%s'</error>", $file));
+                $this->log(sprintf("<error>Plugin config file broken link '%s'</error>", $file));
                 return;
             }
         }
 
-        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-            $this->_output->writeln('<debug>Load plugin config <comment>' . $path . '</comment></debug>');
-        }
-
-        $localPluginConfig = \file_get_contents($path);
-        $localPluginConfig = Yaml::parse($this->applyVariables($localPluginConfig, $magentoRootFolder, $file));
-
-        $this->_pluginConfig = ArrayFunctions::mergeArrays($this->_pluginConfig, $localPluginConfig);
+        $this->logDebug('Load plugin config <comment>' . $path . '</comment>');
+        $localPluginConfigFile = ConfigFile::createFromFile($path);
+        $localPluginConfigFile->applyVariables($magentoRootFolder, $file);
+        $this->_pluginConfig = $localPluginConfigFile->mergeArray($this->_pluginConfig);
     }
 
     /**
@@ -412,5 +392,23 @@ class ConfigurationLoader
     public function getConfigurationLoaderDir()
     {
         return __DIR__;
+    }
+
+    /**
+     * @param string $message
+     */
+    private function logDebug($message)
+    {
+        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
+            $this->log('<debug>' . $message . '</debug>');
+        }
+    }
+
+    /**
+     * @param string $message
+     */
+    private function log($message)
+    {
+        $this->_output->writeln($message);
     }
 }