+++ src/N98/Magento/Command/Developer/ConsoleCommand.php
@@ -27,7 +27,9 @@ class ConsoleCommand extends AbstractMagentoCommand
     {
         $this
             ->setName('dev:console')
-            ->setDescription('Opens PHP interactive shell with initialized Mage::app() <comment>(Experimental)</comment>')
+            ->setDescription(
+                'Opens PHP interactive shell with initialized Mage::app() <comment>(Experimental)</comment>'
+            )
         ;
     }
 
@@ -79,12 +81,8 @@ class ConsoleCommand extends AbstractMagentoCommand
             $magentoVersion = $this->productMeta->getVersion();
 
             $consoleOutput->writeln(
-                '<fg=black;bg=green>Magento '
-                . $magentoVersion
-                . ' '
-                . $edition
-                . ' initialized.</fg=black;bg=green> '
-                . $ok
+                '<fg=black;bg=green>Magento ' . $magentoVersion . ' ' . $edition .
+                ' initialized.</fg=black;bg=green> ' . $ok
             );
         } else {
             $consoleOutput->writeln('<fg=black;bg=yellow>Magento is not initialized.</fg=black;bg=yellow>');