+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 RECENT CHANGES
 ==============
 
+1.1.7
+-----
+
+1.1.6
+-----
+
+* Fix: Fix module loader (reported by Matthias Walter, fixed by TK)
+
+1.1.5
+-----
+* Fix: #172 Base-URL check on IP addresses (TK)
+* Fix: Whitespace and code-style (TK)
+* Feature: Install command: magento-ce-2.0.4 version (TK)
+* Feature: Extract config-loader (TK)
+* Feature: Add dry-run mode for db:dump (TK)
+
 1.1.4
 -----
+* Feature: Install command: magento-ce-2.0.1 and magento-ce-2.0.2 versions (TK)
 * Feature: #101 Porting command: eav:attribute:view (by Robbie Averill)
 * Feature: #120 Porting command: sys:setup:change-version (by Robbie Averill)
 
@@ -29,6 +46,3 @@ RECENT CHANGES
 
 The first n98-magerun2 stable release to power the next-generation
 open source digital commerce platform, Magento 2.0.
-
-
-